requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
sqlalchemy
jinja2
matplotlib
